Auto-resize fix for #6
This change should allow the VNC window to resize to the size of it's
parent div when ever the window is resized, or when the vnc window
itself is resized. Therefore, this fixes #6.

Based off of
https://github.com/jakozaur/noVNC/blob/master/meteor/novnc.js#L56 since
it was the original code that resized the window (and I don't know
meteor or noVnc really well).

Signed-off-by: Colton Wolkins (Ogre) <frostyfrog2@gmail.com>